Do not try to verify focus placement in ClearBrowserDataBrowserTest.
The test was flaky because of a more general bug in the WebUI framework (crbug.com/380737). This latter bug exhibits itself when, during opening a WebUI overlay, the BrowserWindow loses focus -- which happens a lot on build bots running multiple browser tests in parallel.

BUG=379574

Review URL: https://codereview.chromium.org/313843002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@275128 4ff67af0-8c30-449e-8e8b-ad334ec8d88c