Multiple account support in chrome.identity.getAuthToken
This change adds an optional "account" field in the TokenDetails
parameter to getAuthToken. When this field is populated with an object
identifying a valid account signed in on the current profile,
getAuthToken will return a token for the specified account.

When the parameter is not present, getAuthToken behaves as before,
returning tokens for the primary profile user.

BUG=258515

Review URL: https://codereview.chromium.org/293063002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@275173 4ff67af0-8c30-449e-8e8b-ad334ec8d88c