Added BookmarkClient::CanBeEditedByUser.
CanBeEditedByUser indicates if a bookmark can be edited by the user or not.
All of the bookmarks in the bookmark_bar_node, the other_node and the
mobile_node are always editable but some of the extra nodes provided by the
client may be unmodifiable.

Also updated other methods to work only with user-editable nodes, and added
utility helpers to detect nodes that can or can't be edited.

BUG=49598

Review URL: https://codereview.chromium.org/317333004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@275523 4ff67af0-8c30-449e-8e8b-ad334ec8d88c