Separate Settings and Help from History and Extensions when settings-in-a-window is enabled.
With --enable-settings-window, only Settings and Help should be shown in the left nav bar when opening settings or help. Likewise, history and extensions should be the only two nav items when loading those pages.

This is a temporary change to unblock Settings in a Window. Going forward, we will remove the left nav entirely from the Settings window, so the logic for having nav item groups here is only temporary. See issue 294023013 for how this will be done after the branch.

BUG=375425
R=dbeam@chromium.org,sky@chromium.org
TBR=estade@chromium.org
NOTRY=true
# win_chromium_x64_rel flaky failures

Review URL: https://codereview.chromium.org/313363004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276327 4ff67af0-8c30-449e-8e8b-ad334ec8d88c