Add date_synced to PasswordForm.
We need it for the "Remote unsync" project. If user accidentally signed-in in the library then he should be able to clean his passwords. date_synced designates the time the password was downloaded from the sync server. All passwords with (date_synced >= sign-in time) are to be deleted.

BUG=362679

Review URL: https://codereview.chromium.org/316243002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c