Ensure show/hide are properly called on interstitial pages if present.
In my CL for sending Show/Hide to all frames on a page (https://codereview.chromium.org/298283003), I used the RWH for each RenderFrameHost in the tree. This missed the fact that an interstitial page will be overlayed and not part of the frame tree in WebContents.

BUG=381439

Review URL: https://codereview.chromium.org/321253003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276463 4ff67af0-8c30-449e-8e8b-ad334ec8d88c