Make NOTIFICATION_HISTORY_URLS_MODIFIED fired consistently for both typed and non-typed URLs.
The semantics for NOTIFICATION_HISTORY_URLS_MODIFIED was changed as of revision 131865 (https://chromiumcodereview.appspot.com/9852002) so that it would be fired not only for typed URLs, but also for non-typed ones (that is, URLs having |typed_count| == 0).

However, HistoryBackend::AddPagesWithDetails() was not updated to reflect these new semantics. Rectify this omission to make the code less mysterious.

BUG=None

Review URL: https://codereview.chromium.org/323043004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276622 4ff67af0-8c30-449e-8e8b-ad334ec8d88c