Introduce ProfileInvalidationProvider wrapper for InvalidationService
TiclInvalidationService was refactored to no longer depend on Profile.
However, it still is a KeyedService. While this does actually pull in
Profile as a dependency, it implies that TiclInvalidationService should
be considered a per-Profile service, like the other KeyedServices in
Chrome's code base.

This CL introduces ProfileInvalidationProvider, a KeyedService that is a
tiny wrapper around an InvalidationService. Given this wrapper,
InvalidationService (and with it, TiclInvalidationService) no longer have
to be KeyedServices.

BUG=362083
TEST=Updated unit and browser tests

TBR=dcheng (chrome/browser/extensions/api/push_messaging/*)
TBR=davemoore (chrome/browser/profiles/chrome_browser_main_extra_parts_profiles.cc)
TBR=arv (chrome/browser/ui/webui/invalidations_message_handler.cc)

Review URL: https://codereview.chromium.org/327243003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276715 4ff67af0-8c30-449e-8e8b-ad334ec8d88c