Update panel layout immediately after the user has finished resizing a panel.
Update panel layout immediately after the user has finished resizing a panel.
Unfortunately, it is not possible to detect when a resize performed by the
window manager has ended. This CL does the resizing of panels entirely in
Chrome. This is slow and ugly but it is what we used to do in the GTK port.
(panel_drag_gtk.cc)

BUG=371616
TEST=Manual, see bug
R=erg, jianli
TBR=sadrul (For changes in ui/base/x/x11_util.* which have been taken verbatim from https://codereview.chromium.org/264713007/)

Review URL: https://codereview.chromium.org/320713004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276719 4ff67af0-8c30-449e-8e8b-ad334ec8d88c