Change HistoryService::QueryURL to use CancelableTaskTracker
The callback no longer receive an HistoryService::Handle as first parameter
and the caller should use the returned base::CancelableTaskTrasker::TaskId
to cancel a request early.

Instead of passing a CancelableRequestConsumer to the query, the caller
pass a base::CancelableTaskTracker. The lifetime of the callback can be
associated to the request by using base::Passed().

Remove unused types and fix usage of the values passed to the result
callback (check success before using the URLRow or the VisitVector).

BUG=371818
TBR=bauerb,sky,asargent,gavinp,shess,pkasting

Review URL: https://codereview.chromium.org/314293005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@276739 4ff67af0-8c30-449e-8e8b-ad334ec8d88c