Add presubmit script to statically verify ChromeVox using the Closure compiler.
In addition to enforcing that ChromeVox compiles when uploading a change
(on Linux only), this provides a script that can be used to manually
compile some or all of the scripts, optionally based on a set
of modified files.

This adds the necessary externs for compiling ChromeVox.

BUG=371692
R=dtseng@chromium.org

Review URL: https://codereview.chromium.org/323273002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@277585 4ff67af0-8c30-449e-8e8b-ad334ec8d88c