Add new DialogDelegate type for dialogs in the App List
Add a simpler WidgetDelegate for dialogs in the app list. The current
app list dialogs extend DialogDelegate, which can't be customised
enough to achieve the required look for dialogs that open inside the
app list. Also, a lot of functions that WidgetDelegate provides are
not useful for app list dialogs (e.g. OK/cancel buttons, a titlebar,
etc).

For now, only the App Info Dialog will use this dialog type. In a
later commit, all dialogs that can be launched from the app launcher
(including the Uninstall and Create Shortcuts dialogs) will be
modified to use this same dialog style.

BUG=369888

Review URL: https://codereview.chromium.org/306023011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@277691 4ff67af0-8c30-449e-8e8b-ad334ec8d88c