views: Move MenuButton from TextButton to LabelButton.
This also converts the TextButtons in the BookmarkBarView to LabelButtons, since they need to have the same base class as MenuButton does. This also standardizes on fade eliding instead of fade eliding sometimes and ellipsis eliding in other cases.

This patch has the effect that everything on the bookmark bar also renders with GTK+ borders in GTK theme mode on Linux.

BUG=155363

Review URL: https://codereview.chromium.org/298813002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@277878 4ff67af0-8c30-449e-8e8b-ad334ec8d88c