[Mac] Show a warning in the new avatar button/menu if there's an auth error.
This error is shown both in the avatar button, and in the avatar menu. When the 
account for which we're displaying the auth error is clicked in the avatar menu,
we show the Gaia reauth view.

Screenshot: 
https://drive.google.com/file/d/0B1B1Up4p2NRMQW90VmcwcGZQVEU/edit?usp=sharing

I've also fixed a bug where the available text for the account name was 
incorrectly calculated, and eliding overlapped the delete button. Screenshot:
https://drive.google.com/file/d/0B1B1Up4p2NRMaU9DcWdqTVZNMVk/edit?usp=sharing
BUG=311235
TEST= Sign into Chrome with the --new-profile-management flag on. Go to 
accounts.google.com and revoke Chrome's access to that account, from the 
Security tab. Try to bookmark the current page. The avatar button should be
badged like in the attached screenshot.

Review URL: https://codereview.chromium.org/303463010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@277908 4ff67af0-8c30-449e-8e8b-ad334ec8d88c