[Mac] Rebuild app shims when they fail to dyload Chrome Framework.
This also changes app_mode_loader to start Chrome with --app-shim-error
only when it was started by Chrome or if it is the app_list shim. When
started by the user, it launches Chrome with --app-id. This simplifies
how Chrome handles --app-shim-error. It does not need to load the
profile or app extension, and rebuilding the shim does not block
launching the app.

BUG=353047

Review URL: https://codereview.chromium.org/265163006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278126 4ff67af0-8c30-449e-8e8b-ad334ec8d88c