Fix overlap on webstore extensions icon in hidpi
Previously, when the device scale factor >= 1.5, selects the 2x
resource, but then scales it by the DSF. So, for example, at 1.5,
it selects 64x64, and draws it as 96x96 pixels, which causes it to
overlap. Setting the size explicitly seems to fix it, but I'm not
sure if this is the right fix.

R=kalman@chromium.org
BUG=366377
TEST=manual with --force-device-scale-factor={1,1.25,1.5,2}

Review URL: https://codereview.chromium.org/343473003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278325 4ff67af0-8c30-449e-8e8b-ad334ec8d88c