Add back button to chrome://devices
When settings in a window is enabled, add a back button to the devices
page. This helps because the window has no address bar. See bug for
screenshot.

Longer-term, the devices page should show as an overlay in Settings
rather than load as a new page.

RTL is not enabled for this page; see issue 386020.

BUG=384333
R=noamsml@chromium.org,oshima@chromium.org

+oshima@ for icon. Oshima, is theme_resources the right place for this?
+stevenjb@ for browser shenanigans in local_discovery_ui.cc.

Review URL: https://codereview.chromium.org/338423005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278376 4ff67af0-8c30-449e-8e8b-ad334ec8d88c