This CL aims to fix the bug that when webview guest create a SharedWorker, the sctipt file of SharedWorker can not be loaded.
The SharedWorker lives in a separate render process as the webview who creates it. Because  url_request_util::AllowCrossRendererResourceLoad（...） doesn't consider that the render process of the SharedWorker  is a guest, so it denies the load of SharedWorker's script.

So we introduce webview partition id map in ExtensitonRenderState. When loading a SharedWorker script, we will check whether the script file is in the accessible resource of given "Partition ID", and decide whether it can be loaded or not. 

BUG=

Review URL: https://codereview.chromium.org/333313004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278377 4ff67af0-8c30-449e-8e8b-ad334ec8d88c