Rename notification_id() to disambiguate its meaning.
As an intermediate step to disambiguate notifications
at notification UI manager level, we rename notification_id()
to delegate_id() to clarify that it is not an id for
notification in message center.

Also change assignment operator of message center's
notification to protected. Other members are still used by
notification list so they need to be public.

BUG=297867

Review URL: https://codereview.chromium.org/344763002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278593 4ff67af0-8c30-449e-8e8b-ad334ec8d88c