chrome.bluetoothLowEnergy: Introduce connect and disconnect methods.
This CL introduces the connect and disconnect methods to the bluetoothLowEnergy
API, which allow an app to initiate and hold on to the connection to a low
energy device. The API calls internally manage the allocation of an API
resource corresponding to a per app, per device device::BluetoothGattConnection
object.

BUG=381305
TEST=browser_tests --gtest_filter=BluetoothLowEnergyApiTest.*

Review URL: https://codereview.chromium.org/349433002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278669 4ff67af0-8c30-449e-8e8b-ad334ec8d88c