Password bubble: Confirm blacklisting a site if passwords exist.
If a user blacklists a site with existing saved passwords, those
passwords will be deleted. In order to ensure that a user doesn't
accidentally lose data, we prompt them to inform them of the risk, and
ask for confirmation of the decision.

BUG=384155
NOTRY=true

Review URL: https://codereview.chromium.org/347033002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c