Prerender Instant search base url on omnibox focus event.
Prerenders the Instant search base page URL on omnibox focus event provided the 
underlying page is not a search results page and not in incognito profile.      
Prerender request for the current tab is cancelled when a navigation entry is   
committed or when the current tab is deactivated.                               
                                                                                
BUG=384106                                                                      
TEST=Run canary chrome with the following flags:                                
--prefetch-search-results --prerender=enabled --force-fieldtrials="EmbeddedSearch/Group1 prerender_instant_url_on_omnibox_focus:1"

Review URL: https://codereview.chromium.org/338633003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278731 4ff67af0-8c30-449e-8e8b-ad334ec8d88c