Password internals page: notify renderer about logging state on client construction
Currently, ChromePasswordManagerClient informs the PasswordAutofillAgent almost every time about changed availability of the internals page. The only exception is when the client is created. As a result, new tabs opened after a chrome://password-manager-internals tab will only cause browser-side logs.

When the client is created, the agent on the renderer side may not exist yet. So it does not help to send messages to the agent during constructing the client.

Instead, this CL introduces a new "ping" IPC message from the renderer to the browser, which the agent uses to explicitly request logging state update from the client at the time the agent is constructed.

BUG=384269

Review URL: https://codereview.chromium.org/336763002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278738 4ff67af0-8c30-449e-8e8b-ad334ec8d88c