Fix chrome.identity.onSignInChanged regression for sign-outs
This change fixes a regression from r270325, allowing sign-out events
to fire when the --enable-new-profile-management flag is absent, which
is the default.

The previous change added a check to filter out events related to
non-primary accounts, but the primary account ID is not available
during sign-out. Removing the filter fixes the regression.

BUG=376421

Review URL: https://codereview.chromium.org/346913003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278762 4ff67af0-8c30-449e-8e8b-ad334ec8d88c