Move ownership of the ComponentCloudPolicyService to the broker.
The ComponentCloudPolicyService needs to work with the SchemaRegistry,
so it was initially owned by the DeviceLocalAccountPolicyProvider
(which always has a valid reference to the registry). However, this
means that the ComponentCloudPolicyService can't precache any policies
until the session is started.

To enable precaching at enrollment time and at the login screen, the
ComponentCloudPolicyService has been moved to the broker instead.
This allows some code simplifications too, but requires changing the
ownership of the SchemaRegistry for device local accounts.

This CL is just moving code around; a subsequent CL will trigger
the precaching.

TBR=zelidrag@chromium.org
BUG=224596

Review URL: https://codereview.chromium.org/342233005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278847 4ff67af0-8c30-449e-8e8b-ad334ec8d88c