[AiS] Fix ownership issue for image service observers
The Java bridge erroneously did a "delete this" on the observer when
request completion was signaled. That is not necessary, since the
BitmapFetcherService takes ownership of all observers.

As a result of this, there is also no need for the observer to signal
request completion. If clients are interested in taking action on
completion, they can do so in the Observer's destructor.

TBR=sky@chromium.org, tedchoc@chromium.org
BUG=380916
NOTRY=true

Review URL: https://codereview.chromium.org/341083008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@278869 4ff67af0-8c30-449e-8e8b-ad334ec8d88c