Refactor renderer-side script injection. Instead of having two script injection systems (UserScriptScheduler and UserScriptSlave), we have a single system (ScriptInjectionManager).
Introduce abstract ScriptInjection class as an object which knows how to inject
itself and whether or not it is allowed, with two implementations
(UserScriptInjection and ProgrammaticScriptInjection).

Primary benefit:
Combine logic for script injections so that a) it's all in one place, b) when
we ask for permission, we can do so analogously.

Additional benefits:
- Reduce dependency of script injection on ExtensionHelper.
- Eliminate ScriptInjection dependency on UserScriptSlave (or
  ScriptInjectionManager).

BUG=382945
TBR=jschuh@chromium.org (name changes to extension_messages.h, no functional changes)
TBR=jochen@chromium.org (two-line change to chrome/renderer/chrome_content_renderer_client.cc)

Review URL: https://codereview.chromium.org/321993003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@279183 4ff67af0-8c30-449e-8e8b-ad334ec8d88c