Support exposing Mojo services between render frames, render threads, and their respective hosts.
This introduces ServiceRegistry as an abstraction around providing
services to and accessing services from a remote peer. In particular,
this adds peered service registries to RenderProcessHost and
RenderThread, and to RenderFrameHost and RenderFrame - the RenderFrame
setup is implemented using the RenderProcessHost/RenderThread
ServiceRegistry pair.

This replaces the existing WebUI handle setup by adding a webUI
controller service to the frame host registry and a corresponding
request for the webUI controller service to the frame registry.

BUG=386155

Review URL: https://codereview.chromium.org/285333003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@279557 4ff67af0-8c30-449e-8e8b-ad334ec8d88c