[fsp] Add support for MIME type.
This patch adds support for returning mime types with metadata, and using them
for launching file handlers.

Note, that Files app is not able to use MIME types for launching file handlers,
because the MIME types are not passed to Files app. This problem is separately
tracked here: crbug.com/388586.

TEST=unit_tests: *FileSystemProvider*GetMetadata*,
     browser_tests: *FileSystemProvider*MimeType*
BUG=388586
R=benwells@chromium.org, hirono@chromium.org, kinaba@chromium.org

Review URL: https://codereview.chromium.org/352753002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@279972 4ff67af0-8c30-449e-8e8b-ad334ec8d88c