Fix a crash in the window_sizer code which occurs while getting the work area of the window.
The crash occurs on Windows 8 and above in a chrome browser process launched to server ASH before
the ASH shell is initialized. The current behavior in case another browser process launches with a URL
or saved session restore etc is to open those in desktop mode.

That may also not be deseriable given that it makes the chrome browser work in dual mode ASH and desktop.

Will address that in a subsequent patch.

BUG=387598

Review URL: https://codereview.chromium.org/346763010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280174 4ff67af0-8c30-449e-8e8b-ad334ec8d88c