[fsp] Fix parsing a URL in case it is of an isolated type.
Before, the parser class was parsing a virtual_path() member of the passed URL,
which in case of an isolated file system url was of a different format.

This CL resolves this issue by relying on path(), not virtual_path(). The first
one is of the same format for both file system url types.

TEST=unit_test: *FileSystemProvider*Parser_IsolatedURL&
BUG=388963

Review URL: https://codereview.chromium.org/339713007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280248 4ff67af0-8c30-449e-8e8b-ad334ec8d88c