<webview>: Fix issue with "display: none".
If <webview>'s style is set to "display: none", it would remove the
browser plugin or not create it if was "display: none" from the beginning.

When <object> is "display: none", bindings and plugin itself is not
available and attachment will fail after allocating instance ID.
We need to call "-internal-attach" properly. This CL fixes this by:
a) Deferring attachment if plugin bindings is not available. This can be
   identified by simply checking if a binding method on the plugin is a
   typeof 'function'.
b) Calling deferred attachment upon BrowserPlugin initialization.
   This is done by setting an attribute on the <object> ("bindingsavailable"),
   This notifies the mutation observer in the shim which simply does the
   attachment.

This CL also simplifies attachWindowAndSetupEvents() function in the
shim, and renames it to simply "attachWindow" because we do not
set up events in this function anymore.

BUG=387484
Test=In a chrome chrome app, with <webview>.
1) Load a <webview src="..." style="display: none"></webview>,
then set the display to 'block'.
This should load the <webview> properly.
2) Load a <webview src="..."></webview>,
Watch it load ... properly.
Now set its display to 'none':
document.querySelector('webview').style.display = 'none';
Now set its display back to not none:
document.querySelector('webview').style.display = '';
This should load the webview properly.

Review URL: https://codereview.chromium.org/339183014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280440 4ff67af0-8c30-449e-8e8b-ad334ec8d88c