Make ".html" always the default extension for save-as-html feature.
Two changes are involved. One is that putting ".html" in front of
".htm" in the extension list of the file dialog. This way,
Chrome OS file dialog automatically corrects the extension
to ".html" when the user switched mhtml/html type selector.

The other change is to set the default extension to ".html"
on Windows, where we have been using ".htm".

BUG=384204

Review URL: https://codereview.chromium.org/330853003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280513 4ff67af0-8c30-449e-8e8b-ad334ec8d88c