Ignore gnome-keyring deprecation to fix building with Gnome 3.12.
libgnome-keyring is currently used by the password manager, but has been
deprecated in Gnome 3.12. The deprecation warnings cause the build to
fail.

In the long term, we should move to libsecret, which is the recommended
replacement. In the meantime, this change disables the deprecation warnings,
so Chromium can be be built with Gnome 3.12.

BUG=374363,355223
TEST=ninja -C out/Debug chrome

Review URL: https://codereview.chromium.org/292573002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280579 4ff67af0-8c30-449e-8e8b-ad334ec8d88c