Remove NOTREACHED from ContentPasswordManagerDriver::DidLastPageLoadEncounterSSLErrors
ContentPasswordManagerDriver::DidLastPageLoadEncounterSSLErrors currently expects
that at the time it is called, the last committed entry is the active one.

That's not true (http://crbug.com/388246), and while the root cause will be
fixed in a different CL, this one removes a NOTREACHED capturing a wrong
expectation.

BUG=384328

Review URL: https://codereview.chromium.org/356493004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280585 4ff67af0-8c30-449e-8e8b-ad334ec8d88c