Make ChromeSigninClient check if the passed in process id is valid
ChromeSigninClient::IsSigninProcess() checks only if the passed in process is
equal to the signin host id, but the signin host id is initialized to -1 (and
set back if the signin process goes away). This means that if we ever check if
an invalid process is a signin process, the answer is "yes".

Fix this by making ChromeSigninClient::IsSigninProcess check if the process id
is valid first.

Review URL: https://codereview.chromium.org/337263007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@280658 4ff67af0-8c30-449e-8e8b-ad334ec8d88c