Don't use a version script when linking libpdf.so
It appears we used to use a version script when compiling libpdf.so. This was
causing crashes such as the one in the associated bug. The reason given
for using the version map was to hide vendor symbols, but since the PDF code
is all open source now, this shouldn't be an issue.

BUG=386436

Review URL: https://codereview.chromium.org/351773005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@281157 4ff67af0-8c30-449e-8e8b-ad334ec8d88c