Fix flakiness in VerifyFakeInstallAttributesCache.
This test races with a background file write on a WorkerPool thread.
The RunLoop doesn't help because it just spawns another thread in the
pool, instead of flushing the file writing thread.

Fixed with a ScopedAllowIO instead, which mimics the ScopedAllowWait
of the real implementation.

BUG=388148
R=satorux@chromium.org

Review URL: https://codereview.chromium.org/349313008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@281506 4ff67af0-8c30-449e-8e8b-ad334ec8d88c