Replace simple network properties in settings with ONC values
This is step 2 (of many) in converting the network settings to Web UI.

This CL passes the ONC dictionary to the network UI, along with
additional existing non-ONC properties that require more complicated
handling (which will be converted in follow-up CLs).

The only functional change is that previously the string
"Access to this network is protected" was never shown before due to a
typo in JS which was fixed as a renaming side-effect.

BUG=279351
R=pneubeck@chromium.org

Review URL: https://codereview.chromium.org/367453004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@281604 4ff67af0-8c30-449e-8e8b-ad334ec8d88c