Fixed ProfileSyncServiceBookmarkTest on iOS.
This test skips the managed_node because it's not synced. The test
assumed that the managed_node always comes next to the mobile_node but
this is not true on iOS.

This fix changes the test to look for the managed node via
BookmarkClient::CanSyncNode, which is the correct check in the first
place.

BUG=383398

Review URL: https://codereview.chromium.org/376993002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@281847 4ff67af0-8c30-449e-8e8b-ad334ec8d88c