Cut reference from MountedDiskMonitor to DiskMountManager.
For number of reasons:

* The previous code implicitly relied on the fact that the Observer
events are notified to MountedDiskMonitor before VolumeManager
(otherwise VolumeManager::OnDiskEvent misses the hard-unplug event.)
It indeed holds, but it is not a part of explicit contract of base::ObserverList.
We should not rely on the detail.

* It simplifies MountedDiskMonitor (see the number of lines.)

* It simplifies the initialization process of VolumeManager/DiskMountManager,
i.e., removes duplicated RequestMountInfoRefresh calls.
This is desired for Bug 356583.

BUG=356583

Review URL: https://codereview.chromium.org/372853003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@281920 4ff67af0-8c30-449e-8e8b-ad334ec8d88c