Separate the logic of popup alignment and workarea handling as delegate.
MessagePopupCollection contains plenty size of conditions and ifdefs
to work properly with each type of the desktop we have, and some
logic makes side effects on another desktop. This design is unhealty
and adding more conditions sounds incorrect.

Considering this, it would be better to extract platform dependent
parts as a delegate class (PopupAlignmentDelegate) and allow
subclasses to provide platform-specific features.

This design is also beneficial for win-ash, because we had
OS_CHROMEOS layout data and logic which actually means Ash.

BUG=389656
R=stevenjb@chromium.org, dimich@chromium.org
TBR=jamescook@chromium.org, harrym@chromium.org
TEST=message_center_unittests, ash_unittests, some manual checks

Review URL: https://codereview.chromium.org/369573004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@282269 4ff67af0-8c30-449e-8e8b-ad334ec8d88c