Make app list search result mixer more resilient to reordering.
To add more result display types to the app list search results, we plan to
have a large pool of results which will be filtered by display type.

As a displayed result may retain its display index but have a different model
index, it is necessary to handle reordering of the items in an in-place way to
avoid icon flickering.

The current implementation only avoids flickering when the result is in the
same model index which is unsuitable for our purposes.

BUG=349727

Review URL: https://codereview.chromium.org/369693004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@282299 4ff67af0-8c30-449e-8e8b-ad334ec8d88c