Fixes for re-enabling more MSVC level 4 warnings: misc edition #1
This contains fixes for the following sorts of issues:
* Signedness mismatch

The problem here is that using enums to declare bitmasks results in the enum values being signed, when all consumers want to use the values in an unsigned context.  Declaring them as consts allows using the more appropriate uint32 type.

In C++11 we could use "enum class" for this, but C++11 isn't legal yet.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/385443002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@282419 4ff67af0-8c30-449e-8e8b-ad334ec8d88c