Make ChromeAutofillClient a ZoomObserver.
In r273854 we may have removed an indirect pathway used to close open
autofill popups when the associated web_contents undergoes a zoom
change. This CL replaces that by directly registering the
ChromeAutofillClient as a ZoomObserver, allowing it to directly
react to zoom events.

BUG=392412

Review URL: https://codereview.chromium.org/377423002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@282453 4ff67af0-8c30-449e-8e8b-ad334ec8d88c