Fix syncing of system theme on Linux
Makes the "IsSystemThemeDistinctFromDefaultTheme()" detection logic
compatible with non-GTK (ie. Aura) Linux.  It moves the check out of
sync code and into the ThemeService so this kind of bug will be less
likely to recur if the themes logic changes again.

Re-enables an integration test that was disabled during the transition
to Aura.

BUG=304554,392934

Review URL: https://codereview.chromium.org/388563004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@282491 4ff67af0-8c30-449e-8e8b-ad334ec8d88c