Experimental app list: Custom launcher pages have app APIs and style.
Added CustomLauncherPageContents, which is copied from AppWindowContents
and modified to provide launcher pages instead. This correctly
dispatches messages to the extensions system.

The extension inside the launcher page is now "activated" which ensures
that it is styled as a platform app and given the correct APIs by the
extension dispatcher.

BUG=391137
BUG=391202

Review URL: https://codereview.chromium.org/365013003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@282530 4ff67af0-8c30-449e-8e8b-ad334ec8d88c