sync: Re-enable other migration integration tests
Re-enables the migration integration tests that bit-rotted while the
tests were temporarily (and accidentally) disabled.

Make the BackendMigrator accept the DataTypeManager::PARTIAL_SUCCESS
status as if it were DataTypeManager::OK.  This is necessary because
the Managed User Settings type will always be "unready", so the
DataTypeManager will never make it to an "OK" state.

The second step was to avoid migrating the Managed User Settings type.
Since it is unready, it will never be enabled and therefore never be
migrated.  Trying to migrate it just confuses the test harness, as it
will wait indefinitely for a migration that never occurs.

BUG=392989

Review URL: https://codereview.chromium.org/386133002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@283129 4ff67af0-8c30-449e-8e8b-ad334ec8d88c