mac: Load the system hotkeys after launch.
Shortly after launch, the system hotkeys are loaded and parsed. If a hotkey is
reserved by the system, it is not passed to the renderer. This allows system
hotkeys like (cmd + `) to work even if a flash plugin is selected.

Add a histogram to ensure that the system hotkey plist is being correctly
loaded and parsed.

BUG=383558
TEST=Open 2 Chrome windows. Navigate one to www.twitch.tv. The site should
include a flash plugin that automatically starts playing a video. Select the
flash plugin. The hotkey combination (cmd + `) should switch between the open
windows. The hotkey combination (cmd + L) should have no effect (it is a Chrome
hotkey, not a browser hotkey).

Review URL: https://codereview.chromium.org/370293004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@283405 4ff67af0-8c30-449e-8e8b-ad334ec8d88c