[fsp] Refactor FakeProvidedFileSystem to be more flexible.
Before, the FakeProvidedFileSystem had some hardcoded responses. However, since
FileStreamWriter is coming soon, we will need to verify if the file on the
file system is changed correctly.

For that, FakeProvidedFileSystem has been refactored, so it is possible to add
and get fake files and verify their values.

TEST=unit_tests: *FileSystemProvider*
BUG=391362

Review URL: https://codereview.chromium.org/393233002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@283538 4ff67af0-8c30-449e-8e8b-ad334ec8d88c