[Profiles] With new-profile-management all new profiles should have default avatars.
In the "delete the only active profile and create a new one" case, the newly
created profile should use the placeholder avatar, and not one of the custom ones. 
The same should happen for the first run default profile.

BUG=NONE
TEST=Start Chrome with --new-profile-management. Delete all the profiles you have.
The newly created profile should use the default guest-looking avatar (which you
can check in the avatar bubble).

Review URL: https://codereview.chromium.org/392313002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@283635 4ff67af0-8c30-449e-8e8b-ad334ec8d88c