Remove disable-pnacl-install flag.
This was added a while back to be able to quickly turn off
component downloads for PNaCl if it got out of hand (via
flipping the flag).

It doesn't seem so necessary anymore, so remove it as part
of the chrome flags cleanup effort.


BUG=357762

Review URL: https://codereview.chromium.org/398213002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@283698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c