Omnibox: Don't Autocomplete Previously-Issued Search Queries
If I type an exact query that I issued before, SearchProvider shouldn't
offer an extension to the autocompletion, even if that I issued that longer
query more recently.  This is counter to user's expectations: if I issued
the query before and type it again and hit enter, I should get that query.

I do not plan to bother running an experiment for this change.

BUG=246411

Review URL: https://codereview.chromium.org/397723003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@283836 4ff67af0-8c30-449e-8e8b-ad334ec8d88c