Add app icon masp to notification API
Add an app icon mask (16*16, or 32*32 for retina display) on the bottom right corner of notifications to indicate source of the notification. The app icon is displayed in a way that takes only the alpha channel of the image passed in, masks it with white foreground and grey background.

BUG=284592

Review URL: https://codereview.chromium.org/308053004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@284586 4ff67af0-8c30-449e-8e8b-ad334ec8d88c