Provide an option to launch the date and time preferences.
When an X.509 certificate appears invalid because it is expired or not yet 
valid, but we have reason to believe the clock is wrong, help users find out 
how to fix their clock.

A future CL will run a check for the time at startup (and periodically
after?) and raise an infobar with similar functionality, as well.

BUG=349653

Review URL: https://codereview.chromium.org/397713007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@284638 4ff67af0-8c30-449e-8e8b-ad334ec8d88c