Hook up .d files and outputs to grit
Pull grit 171 to get the new deps fixes required to implement this.

This removes the call to grit for querying inputs completely, and enables depfile writing to enable automatic rebuilds when things are out-of-date.

Adds a new outputs variable that bypasses querying grit for outputs. This will assert that the given files exist in grit. This version allows outputs to be empty in which case it will revert to the old dynamic behavior. I'm going to land this with just the worst cases fixed to make landing easier. Then followup with converting the rest of the grit targets and removing the dynamic option.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/407653003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@284705 4ff67af0-8c30-449e-8e8b-ad334ec8d88c