Content Verification: Don't access UI-thread objects on the IO thread
We ask to create ContentVerifyJob's on the IO thread as we read files
from extension directories, but we are accidentally accessing state that
lives on the UI thread (eg the ExtensionRegistry) while deciding whether
to do that or not.

This changes things so that as we discover extensions loading up, we
copy the data we need to make those decisions to an object on the IO
thread.

Also fix a bug that was keeping BOOTSTRAP mode from working, and one or
two other minor issues I found while I was staring at things.

BUG=395873,392216

Review URL: https://codereview.chromium.org/407043002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@284804 4ff67af0-8c30-449e-8e8b-ad334ec8d88c