android: add ThumbnailCache
The purpose of the thumbnail cache is to save resources and use 
static image instead of live pages when contents of a tab are
needed for the browser UI.

There are three forms of caching:
  - High-res
  - Low-res
  - On-disk

If a high-res thumbnail is not available, then a low-res thumbnail
is used while the on-disk version (always high-res) is being read.

Eviction from the high/low-res caches is determined by a
priority-ordered list of tab ids that is updated by the user.

Compression of the thumbnail bitmaps are supported through
third_party/android_opengl/etc1.  The raw bitmaps are discarded
once a compressed version is available.  Compressed bitmaps are
written to disk and uploaded as UIResource.  Once both of tasks
are complete, the compressed bitmap is also discarded.

depends on: https://codereview.chromium.org/287593002/

android= https://chrome-internal-review.googlesource.com/#/c/160907/

BUG=341406, 357740

Review URL: https://codereview.chromium.org/262543002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@284861 4ff67af0-8c30-449e-8e8b-ad334ec8d88c