GN: Add //ui/base/ime, fix up printing on mac
This adds the //ui/base/ime target to the GN build, which is used
everywhere except for ios. In the gyp build this isn't a real target,
it's a set of properties that are included into the ui_base target which
is the best you can do in GYP since you can't link a static library into
a component without dropping some symbols. In GN, we can define this as
a source_set with its own dependencies and settings.

This also fixes up some miscellaneous mac issues.

R=brettw@chromium.org,ben@chromium.org

Review URL: https://codereview.chromium.org/389083002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@284981 4ff67af0-8c30-449e-8e8b-ad334ec8d88c