Fix races in DeviceLocalAccountTest.Extensions*
The DeviceLocalAccountTest.Extensions* assumed that extensions get copied
to an account's cache after successful installation. In reality,
extensions are downloaded to the cache at first and then removed again if
installation fails. The removal happens asynchronously on a background
thread. This led to a race where tests would try to verify that an
extension does not exist in a cache without waiting for the removal
process to finish first.

This CL switches the tests from verifying that an extension has been
removed from disk to checking whether the cache entry has been removed,
which happens synchronously on the UI thread.

BUG=393208
TEST=DeviceLocalAccountTest.Extensions*

Review URL: https://codereview.chromium.org/389313003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285070 4ff67af0-8c30-449e-8e8b-ad334ec8d88c