Change semantics from newProfileManagement() to accountConsistency().
Account Consistency should enable the Accounts Reconcilor, the multi-accounts UI in the user menu, and inform GAIA to show Manager Accounts not Add Account & Sign Out. This CL also should disable the promo card that leads a user from NewProfileManager to EnableAccountConsistency.

Please note the first message of the CL contains a full list of all references to IsNewProfileManagement and why each was or was not included in this change.

BUG=391497
TEST=Setting the AccountConsistency flag should control all of (and
only) the synchronization of the token service and the cookie jar. All
Mirror-UI should be behind newAvatarMenu and NewProfileManagement.

Review URL: https://codereview.chromium.org/372033002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c