Remove some very old-file-format code from safe browsing store.
In 2010 safe-browsing storage was rewritten to use a binary format
instead of SQLite, with support for SQLite removed in 2011 (r74253).
Remove the code which detects and histograms this, as there is no
possibility that anything will be done with that data.

Additionally, remove code to delete even older SQLite files left from
the original safe-browsing implementation.  This code is firing a few
thousand times per month, but it is hard to imagine any scenario where
profiles that continue to contain this data are not horribly broken.

BUG=none
R=asvitkine@chromium.org, feng@chromium.org

Review URL: https://codereview.chromium.org/404803006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285389 4ff67af0-8c30-449e-8e8b-ad334ec8d88c