chromeos: Refactor the code for showing the network UI.
This patch moves some code out of //ash into //chromeos so that it can be used
in other components (e.g. athena). NetworkListView is introduced for representing
the list of available networks to the user. It can query information through the
NetworkListDelegate. NetworkInfo is used to contain the necessary information
about a network.

BUG=387199
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/417793003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285411 4ff67af0-8c30-449e-8e8b-ad334ec8d88c