Resubmit: Open the WrenchMenu on mouseover when dragging a browser action
Original Patch: https://codereview.chromium.org/399143004
Reverted In: https://codereview.chromium.org/416903002/
This was reverted because the test was flaky on Vista. Resubmit with a fix.

------------------------

Open the WrenchMenu on mouseover when dragging a browser action

Enable dragging from the main container to the overflow container by opening the
wrench menu when hovering over the button while dragging a browser action. This
also requires keeping the menu open while the drag is happening.

Note: this doesn't fix drag calculations for the overflow menu.
In an effort to keep these patches small, that will come later.

BUG=393038

Review URL: https://codereview.chromium.org/413323002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285634 4ff67af0-8c30-449e-8e8b-ad334ec8d88c