Update Confusing Log Message
The chrome.notifications.update will reasonably not update notifications that
do not exist. Previously, the console would log that the update failed to occur
since there was no update. This gives the wrong impression of an error.

This delta relaxes the wording to "not updated".

BUG=396881

Review URL: https://codereview.chromium.org/416913002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285649 4ff67af0-8c30-449e-8e8b-ad334ec8d88c